+++ exonum-light-client/CHANGELOG.md
@@ -15,7 +15,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
-###
+### Added
+- `ExonumClient#getServiceInfoByName(String)` to retrieve a service id by its
+  name and `ExonumClient#getServiceInfoList` to retrieve the list of all
+  started services - their names and ids.
+
+### Changed
 - `TransactionResponse#getExecutionResult` now returns `ExecutionStatus`
   instead of `TransactionResult`. (#1244)